Move statics to separate file
